Tidy the file GenMesh.h up
The problem was to determine which of the branches

	develop
	main

was closer to the versions of the profiler released while in Japan,
so I checked every source file. The branch I was looking for is
finally develop.

The branch main still has stuff of the inner mesh to be removed. In this
commit I removed inner mesh headers in GenMesh.h.